Move notfound to containers and proplerly assign proptype rules
