+++ package.json
@@ -41,7 +41,7 @@
     "budo": "^6.0.2",
     "envify": "^3.4.0",
     "eslint": "^1.7.2",
-    "eslint-plugin-react": "^3.6.3",
+    "eslint-plugin-react": "^3.10.0",
     "smokestack": "^3.4.1",
     "tap-status": "^1.0.1",
     "tape": "^4.2.2"
